Workaround: Ignore standby state for background image
During the standby state, robots are standing in the same
place for a long time. This is bad for computing the
background image. This is just a hack: The first two
minutes are simply skipped when computing the background
image. This does not consider how long the standby state
actually was and whether there also were ones in the
middle of a half due to timeouts.